create CardReports with props and refactory tabs pages
