+++ witch_oracle/src/app/dashboard/tabs/cubes/page.tsx
@@ -7,7 +7,7 @@ export default function EverySection({ from = '' }) {
 
     return <div>
 
-        <CardsReport from="attack"></CardsReport>
+        <CardsReport from="cubes"></CardsReport>
 
     </div>;
